Import of PLU/Gearman-Driver-0.01999_01 from CPAN.
gitpan-cpan-distribution: Gearman-Driver
gitpan-cpan-version:      0.01999_01
gitpan-cpan-path:         PLU/Gearman-Driver-0.01999_01.tar.gz
gitpan-cpan-author:       PLU
gitpan-cpan-maturity:     developer

+++ lib/Gearman/Driver/Job.pm
@@ -3,7 +3,6 @@ package Gearman::Driver::Job;
 use Moose;
 use Gearman::Driver::Adaptor;
 use POE qw(Wheel::Run);
-use IO::Socket;
 
 =head1 NAME
 
@@ -59,28 +58,15 @@ has 'name' => (
     required => 1,
 );
 
-=head2 method
+=head2 methods
 
-Code reference which is run called by L<Gearman::XS::Worker>.
-Actually it's not called directly by it, but in a wrapped coderef.
-
-=cut
-
-has 'method' => (
-    is       => 'rw',
-    isa      => 'CodeRef',
-    required => 1,
-);
-
-=head2 worker
-
-Reference to the worker object.
+ArrayRef of L<Gearman::Driver::Job::Method> objects.
 
 =cut
 
-has 'worker' => (
+has 'methods' => (
     is       => 'rw',
-    isa      => 'Any',
+    isa      => 'ArrayRef[Gearman::Driver::Job::Method]',
     required => 1,
 );
 
@@ -107,43 +93,6 @@ has 'min_processes' => (
     default  => 1,
     is       => 'rw',
     isa      => 'Int',
-    required => 1,
-);
-
-=head2 encode
-
-This may be set to a method name which is implemented in the worker
-class or any subclass. If the method is not available, it will fail.
-The returned value of the job method is passed to this method and
-the return value of this method is sent back to the Gearman server.
-
-See also: L<Gearman::Driver::Worker/Encode>.
-
-=cut
-
-has 'encode' => (
-    default  => '',
-    is       => 'rw',
-    isa      => 'Str',
-    required => 1,
-);
-
-=head2 decode
-
-This may be set to a method name which is implemented in the worker
-class or any subclass. If the method is not available, it will fail.
-The workload from L<Gearman::XS::Job> is passed to this method and
-the return value is passed as argument C<$workload> to the job
-method.
-
-See also: L<Gearman::Driver::Worker/Decode>.
-
-=cut
-
-has 'decode' => (
-    default  => '',
-    is       => 'rw',
-    isa      => 'Str',
     required => 1,
 );
 
@@ -209,51 +158,28 @@ has 'session' => (
     isa => 'POE::Session',
 );
 
-has 'client' => (
-    builder => '_build_client',
-    is      => 'ro',
-    isa     => 'IO::Socket::UNIX',
-    lazy    => 1,
-);
-
 =head2 lastrun
 
 Each time this job is called it stores C<time()> in this attribute.
 
-This depends on L<Gearman::Driver/extended_status>.
-
 =cut
 
 has 'lastrun' => (
+    default => 0,
     is      => 'rw',
     isa     => 'Int',
-    trigger => sub {
-        my ( $self, $value, $old_value ) = @_;
-        $old_value ||= 0;
-        return if $value == $old_value;
-        my $sock = $self->client;
-        print $sock $self->name . " lastrun $value\n";
-    }
 );
 
 =head2 lasterror
 
 Each time this job failed it stores C<time()> in this attribute.
 
-This depends on L<Gearman::Driver/extended_status>.
-
 =cut
 
 has 'lasterror' => (
+    default => 0,
     is      => 'rw',
     isa     => 'Int',
-    trigger => sub {
-        my ( $self, $value, $old_value ) = @_;
-        $old_value ||= 0;
-        return if $value == $old_value;
-        my $sock = $self->client;
-        print $sock $self->name . " lasterror $value\n";
-    }
 );
 
 =head2 lasterror_msg
@@ -261,20 +187,24 @@ has 'lasterror' => (
 Each time this job failed it stores the error message in this
 attribute.
 
-This depends on L<Gearman::Driver/extended_status>.
-
 =cut
 
 has 'lasterror_msg' => (
+    default => '',
     is      => 'rw',
     isa     => 'Str',
-    trigger => sub {
-        my ( $self, $value, $old_value ) = @_;
-        $old_value ||= '';
-        return if $value eq $old_value;
-        my $sock = $self->client;
-        print $sock $self->name . " lasterror_msg $value\n";
-    }
+);
+
+=head2 worker
+
+Reference to the worker object.
+
+=cut
+
+has 'worker' => (
+    is       => 'rw',
+    isa      => 'Any',
+    required => 1,
 );
 
 =head1 METHODS
@@ -306,48 +236,9 @@ sub BUILD {
 
     $self->{gearman} = Gearman::Driver::Adaptor->new( server => $self->driver->server );
 
-    my $extended_status = $self->driver->extended_status;
-
-    my $decoder = sub { shift };
-    my $encoder = sub { shift };
-
-    if ( my $decoder_method = $self->decode ) {
-        $decoder = sub { return $self->worker->$decoder_method(@_) };
+    foreach my $method ( @{ $self->methods } ) {
+        $self->gearman->add_function( $method->name => $method->wrapper );
     }
-    if ( my $encoder_method = $self->encode ) {
-        $encoder = sub { return $self->worker->$encoder_method(@_) };
-    }
-
-    my $wrapper = sub {
-        my ($job) = @_;
-
-        my @args = ($job);
-
-        push @args, $decoder->( $job->workload );
-
-        $self->worker->begin(@args);
-
-        my $error;
-        my $result;
-        eval {
-            $result = $self->method->( $self->worker, @args );
-        };
-        if ($@) {
-            $error = $@;
-            $self->lasterror(time) if $extended_status;
-            $self->lasterror_msg($error) if $extended_status;
-        }
-
-        $self->lastrun(time) if $extended_status;
-
-        $self->worker->end(@args);
-
-        die $error if $error;
-
-        return $encoder->($result);
-    };
-
-    $self->gearman->add_function( $self->name => $wrapper );
 
     $self->{session} = POE::Session->create(
         object_states => [
@@ -362,15 +253,6 @@ sub BUILD {
             }
         ]
     );
-}
-
-sub _build_client {
-    my ($self) = @_;
-    IO::Socket::UNIX->new(
-        Peer    => $self->driver->cc_socket,
-        Type    => SOCK_STREAM,
-        Timeout => 2,
-    ) or die $@;
 }
 
 sub _start {
@@ -406,6 +288,7 @@ sub _add_process {
 sub _remove_process {
     my ( $self, $kernel, $heap ) = @_[ OBJECT, KERNEL, HEAP ];
     my ($pid) = ( $self->get_pids )[0];
+    return unless $pid;
     my $process = $self->delete_process($pid);
     $process->kill();
     $self->log->info( sprintf '(%d) [%s] Process killed', $process->PID, $self->name );
@@ -414,7 +297,9 @@ sub _remove_process {
 sub _on_process_stdout {
     my ( $self, $heap, $stdout, $wid ) = @_[ OBJECT, HEAP, ARG0, ARG1 ];
     my $process = $heap->{wheels}{$wid};
-    $self->log->info( sprintf '(%d) [%s] STDOUT: %s', $process->PID, $self->name, $stdout );
+    my ( $attr, $value ) = $stdout =~ /^(\w+) (.*?)$/;
+    return if !defined $attr || !defined $value;
+    $self->$attr($value) if $self->can($attr);
 }
 
 sub _on_process_stderr {
@@ -472,6 +357,8 @@ See L<Gearman::Driver>.
 =item * L<Gearman::Driver::Console::Basic>
 
 =item * L<Gearman::Driver::Console::Client>
+
+=item * L<Gearman::Driver::Job::Method>
 
 =item * L<Gearman::Driver::Loader>
